FIX Do not fail on errors in *.mk files anymore - except in interactive mode
If a syntax error in <tt>main.mk</tt> or other <tt>*.mk</tt> files occurs,
Check_MK used to abort any operation. This is nasty when for example a new
variable that has been introduced in a new version is not known in an older
version of Check_MK and thus after a version downgrade the configuration
cannot be activated any more.

The behaviour has now changed so that in case of such an error Check_MK will only
abort in <i>interactive</i> mode, i.e. if the standard output is a terminal.
That means that WATO will always try to activate Changes. You can force the
old behaviour by adding the new option <tt>--interactive</tt>.

Conflicts:
	ChangeLog
	modules/check_mk.py