check_dns: It is now possible to use the local dns server in wato configuration
Befor this change, it was only possible to define the ip of a dns server or use the the ip addres of the host as dns server. Now it's to omit this confiuration and the check is done against the dns servers configured on the monitoring servers.

This is usefull if you e.g. want check web domain names.