FIX fc_port: do not inventorize if brocade fibre channel mib is also supported on the device
Some devices on the market have shown to support both the brocade fibre channel mib, and
the FCMGMT mib. In these cases two services were created for each fibre channel port, one for
the brocade_fcport check and one for the fc_port check.
With this fix, the fc_port does not inventorize if the brocade fibre channel mib is supported.