#0647 FIX printer_input, printer_output: multiple fixes to algorithm for state determination and inventory function
The inventory function now considers the description of the unit if its name is not configured. This should lead to
more and more meaningful items. Still, all units with empty name and description will be merged into one single
service with an empty item name.

The algorithm to determine the states of a unit was plain wrong, so that a nonsense collection of states were given.
This has been corrected. Still several states may be derived, but they should now correctly describe the operational
state of the unit.

The inventory function also used the same algorithm to inventorize only units with meaningful states. Therefore some
changes may be expected in the inventory of a printer as well.