#1157 FIX Fixed SMS plugin on at least debian (distrs which have no sendsms/smssend)
Some distros do not provide the sendsms/smssend script with the smstools package.
In this cases the script for sending SMS did not work. This commit adds a dedection
when smssend/sendsms are not available whether the spool directory at
/var/spool/sms/outgoing is available or not. When it is available, the SMS notification
plugin drops a file to this directory which should then be processed by smsd.