#1394 HTML notifications have a new content field for debugging variables
When configuring notifications and customizing templates then you might often
wonder, which variables are present and which values they have. In the HTML
email notifications you can now enable the new field <i>Complete variable list</i>
in the <i>Information to be displayed in the email body<i/>. This adds a table
with all avaiable variables and their current settings in the notification
in question.