#1112 esx_vsphere_hostsystem.mem_usage_cluster: allows to monitor total RAM usage of all nodes in a cluster
This check can not be inventorized. You need to create a manual check and assign it to a cluster.
The memory information of each cluster node is added to the total memory pool.
You are able to set limits for the total RAM usage depending on the number of cluster nodes.