#1401 FIX Display options in views are now again persistent
When you change a "Display Option" in a view - such as the <i>Time stamp format</i>, then
this option was only valid whilst you were visiting that view. This was due to a bug, which
was introduced in Januar 2014. This has been fixed now. If you are visiting a view again
after changing display options, then these will be persisted forever. Note: the persistance
is on a per-user-base and per-view-base.