#1405 Checkbox for settings downtimes on the hosts of the selected services
In the Multisite command box for setting downtimes on services there is now
a new checkbox called <i>Schedule downtimes on the affected hosts instead
of their services</i>.  When you check this then the scheduled downtimes
are set on the hosts of the selected services rather then on the services
themselves. Note: When a host is in scheduled downtime then all its services
are considered to be in a scheduled downtime, as well. There is no need to
set a downtime on these.