#1483 FIX Savely replace illegal vertical bars in check plugin output
If a check plugin has a vertical bar in its plugin output then the core chopped
off the remaining text and interpreted that is performance values. That was
for example the case when monitoring SAP R/3 jobs. This has now been fixed
by automatically replacing the bar by the Unicode character LIGHT VERTICAL BAR
(U+2758).