#1495 Most WATO tables can now be sorted (where useful)
Most of the data tables in WATO can now be used. Some of the
tables which have a defined order, like the rules within a
ruleset can not be sorted using the new sort mechanism.

At the moment the host table is not sortable, but this will
be added soon.