#1534 FIX Fixed filtering views in distributed setup lead to empty views
This fixes an error where filtering a view leads into an empty view, because
the enforced site filter has the first site preselected, which is not always
the site of the data shown in the view. Now we use the optional site filter
for most views to cleanup this error.