#1585 FIX Dashboard: Fixed mass client CPU load consumption when making graph dashlets too small
When PNP graph dashlets were too small, the PNP graph resizing code could get into
a situation where continous HTTP requests were made to get a smaller PNP graph where
it was not possible to get a smaller one. This resulted in a lot of CPU load by the
browser on the client host.