#1715 Output icon information in CSV/JSON/Python export of views
When exporting views to JSON, Python or CSV the column with the icons
is now no longer empty but is a space separated list of all icon names
of the host/sevice in question - without path information or the suffixes
of the file type. That way you now can get additional information about
the objects.