#1791 FIX Fix problem where many bogus RRD files for Check_MK service would be created
The problem appeared when using CMC and during check execution some general problem
appeared like <tt>UNKNOWN - global name 'netsnmp' is not defined</tt>. Then in the
<i>next</i> check done by the same helper this message would appear again because
one variable was not cleared. That garbled the protocol with the microcore and the
later took the output of the passive checks as performance data. This has been
fixed.